// src/app/model/formBuild.ts
export class InitializrForm {
  project: string = "";
  type: string = "";
  language: string = "";
  bootVersion: string = "";
  domain: string = "";
  groupId: string = "";
  artifactId: string = "";
  name: string = "";
  description: string = "";
  packageName: string = "";
  packaging: string = "";
  javaVersion: string = "";
  dependencies: string[] = [];
  baseDir: string = "";
  version: string = "";
}